+++ plugin-build/src/test/kotlin/androidx/sqlite/SQLiteDriver.kt
@@ -1,0 +1,3 @@
+package androidx.sqlite
+
+interface SQLiteDriver